requirements.txt
torch
torchvision
python-chess~=0.31.4
numpy
setuptools
requests